api/tvlp: remove start_time in-query parameter
