+++ tests/aat/api/v1/client/api/tvlp_api.py
@@ -719,7 +719,6 @@ class TVLPApi(object):
 
         :param async_req bool
         :param str id: Unique resource identifier (required)
-        :param datetime time: The ISO8601-formatted date and time to start profile replay
         :param TvlpStartConfiguration start: TVLP Start parameters
         :return: TvlpResult
                  If the method is called asynchronously,
@@ -743,14 +742,13 @@ class TVLPApi(object):
 
         :param async_req bool
         :param str id: Unique resource identifier (required)
-        :param datetime time: The ISO8601-formatted date and time to start profile replay
         :param TvlpStartConfiguration start: TVLP Start parameters
         :return: TvlpResult
                  If the method is called asynchronously,
                  returns the request thread.
         """
 
-        all_params = ['id', 'time', 'start']  # noqa: E501
+        all_params = ['id', 'start']  # noqa: E501
         all_params.append('async_req')
         all_params.append('_return_http_data_only')
         all_params.append('_preload_content')
@@ -777,8 +775,6 @@ class TVLPApi(object):
             path_params['id'] = params['id']  # noqa: E501
 
         query_params = []
-        if 'time' in params:
-            query_params.append(('time', params['time']))  # noqa: E501
 
         header_params = {}
