+++ tests/aat/api/v1/docs/TVLPApi.md
@@ -340,7 +340,7 @@ No authorization required
 [[Back to top]](#) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to Model list]](../README.md#documentation-for-models) [[Back to README]](../README.md)
 
 # **start_tvlp_configuration**
-> TvlpResult start_tvlp_configuration(id, time=time, start=start)
+> TvlpResult start_tvlp_configuration(id, start=start)
 
 Start a TVLP configuration
 
@@ -357,12 +357,11 @@ from pprint import pprint
 # create an instance of the API class
 api_instance = client.TVLPApi()
 id = 'id_example' # str | Unique resource identifier
-time = '2013-10-20T19:20:30+01:00' # datetime | The ISO8601-formatted date and time to start profile replay (optional)
 start = client.TvlpStartConfiguration() # TvlpStartConfiguration | TVLP Start parameters (optional)
 
 try:
     # Start a TVLP configuration
-    api_response = api_instance.start_tvlp_configuration(id, time=time, start=start)
+    api_response = api_instance.start_tvlp_configuration(id, start=start)
     pprint(api_response)
 except ApiException as e:
     print("Exception when calling TVLPApi->start_tvlp_configuration: %s\n" % e)
@@ -373,7 +372,6 @@ except ApiException as e:
 Name | Type | Description  | Notes
 ------------- | ------------- | ------------- | -------------
  **id** | **str**| Unique resource identifier | 
- **time** | **datetime**| The ISO8601-formatted date and time to start profile replay | [optional] 
  **start** | [**TvlpStartConfiguration**](TvlpStartConfiguration.md)| TVLP Start parameters | [optional] 
 
 ### Return type
