+++ tests/aat/api/v1/docs/TvlpStartConfiguration.md
@@ -3,6 +3,7 @@
 ## Properties
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
+**start_time** | **datetime** | The ISO8601-formatted date and time to start profile replay | [optional] 
 **cpu** | [**TvlpStartSeriesConfiguration**](TvlpStartSeriesConfiguration.md) |  | [optional] 
 **memory** | [**TvlpStartSeriesConfiguration**](TvlpStartSeriesConfiguration.md) |  | [optional] 
 **block** | [**TvlpStartSeriesConfiguration**](TvlpStartSeriesConfiguration.md) |  | [optional]
