+++ tests/aat/api/v1/docs/TvlpStartSeriesConfiguration.md
@@ -5,7 +5,6 @@ Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
 **load_scale** | **float** | The scale multiplier for load parameters of generators | [optional] [default to 1.0]
 **time_scale** | **float** | The scale multiplier for the length of each profile entry | [optional] [default to 1.0]
-**start_time** | **datetime** | The ISO8601-formatted date and time to start profile replay | [optional] 
 **dynamic_results** | [**DynamicResultsConfig**](DynamicResultsConfig.md) |  | [optional] 
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
